requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
black==25.1.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
croniter==6.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
Flask==3.1.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
numpy==2.2.6
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pika==1.3.2
platformdirs==4.4.0
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==7.0.0
requests==2.32.4
rq==2.6.0
schedule==1.2.2
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.47.2
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
